Policys
Se avanza en la configuración de los policyes, pendiente validación de front.

+++ app/Http/Controllers/BusinessIdeasController.php
@@ -27,7 +27,7 @@ class BusinessIdeasController extends Controller
         $user       = auth()->user();
         $business   = $user->business()->first();
         $ideas      = BusinessIdeas::where('business_id','=',$business->id)->Paginate(10);
-        //$this->authorize('viewAny',[BusinessIdeas::class, $ideas]);
+        $this->authorize('viewAny',[BusinessIdeas::class, $ideas]);
 
         return view('BusinessIdeas.index', compact('ideas','business'));
     }
@@ -42,7 +42,8 @@ class BusinessIdeasController extends Controller
         $user           = auth()->user();
         $user_business  = $user->business()->first();
         $idea           = BusinessIdeas::find($id);
-        //$this->authorize('view',[BusinessIdeas::class, $idea]);
+       
+        $this->authorize('view',[BusinessIdeas::class, $idea]);
 
         return view('BusinessIdeas.show', compact('idea','user_business'));
     }
@@ -56,7 +57,7 @@ class BusinessIdeasController extends Controller
         $user = auth()->user();
         $user_business = $user->business()->first();
         $idea = New BusinessIdeas;
-        //$this->authorize('create', [BusinessIdeas::class, $idea]);
+        $this->authorize('create', [BusinessIdeas::class, $idea]);
 
         return view('BusinessIdeas.create', compact('user_business'));
     }
@@ -84,12 +85,12 @@ class BusinessIdeasController extends Controller
         $business_idea->money           = $request->get('money');
         $business_idea->condition       = $request->get('condition');
 
-        //$this->authorize('create', [BusinessIdeas::class, $business_idea]);
+        $this->authorize('create', [BusinessIdeas::class, $business_idea]);
 
         if($business_idea->save()){
             // ? Crea un nuevo objeto de la clase Notification y ejecuta el método newBusinesIdea para crear la notificación de la idea
-            $notification = New NotificationController;
-            $notification->newBusinessIdea($business_idea);
+            //$notification = New NotificationController;
+            //$notification->newBusinessIdea($business_idea);
             return redirect()->route('business-ideas.index')->with('success', 'La idea fue creada correctamente');
         } else{
             return redirect()->route('business-ideas.create')->with('success', 'Hubo un error al momento de crear la idea');
@@ -107,7 +108,8 @@ class BusinessIdeasController extends Controller
         $user           = auth()->user();
         $user_business  = $user->business()->first();
         $idea           = BusinessIdeas::find($id);
-        //$this->authorize('update', [BusinessIdeas::class, $idea]);
+        
+        $this->authorize('update', [BusinessIdeas::class, $idea]);
 
         return view('BusinessIdeas.edit', compact('idea','user_business'));
     }
@@ -131,7 +133,8 @@ class BusinessIdeasController extends Controller
         $idea->have_money      = $request->get('have_money');
         $idea->money           = $request->get('money');
         $idea->condition       = $request->get('condition');
-        //$this->authorize('update', [BusinessIdeas::class, $idea]);
+
+        $this->authorize('update', [BusinessIdeas::class, $idea]);
 
         if($idea->save()){
             return redirect()->route('business-ideas.index')->with('success', 'La idea fue editada correctamente');;
@@ -147,7 +150,11 @@ class BusinessIdeasController extends Controller
      */
     public function destroy($id)
     {
-        //$this->authorize('delete', [BusinessIdeas::class, $idea]);
+        $user = auth()->user();
+        $business = $user->business()->first();
+        $idea = BusinessIdeas::find($id);
+
+        $this->authorize('delete', [BusinessIdeas::class, $idea]);
         $idea = BusinessIdeas::find($id)->delete();
 
         return redirect()->route('business-ideas.index')
